+++ app/src/main/java/io/github/wulkanowy/data/db/entities/Notification.kt
@@ -19,6 +19,7 @@ data class Notification(
 
     val type: NotificationType,
 
+    @ColumnInfo(defaultValue = "")
     val destination: Destination,
 
     val date: LocalDateTime,
